refactor: clean up tests
